refactor: move the routes into each sub-modules
